requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
pydantic-extra-types
phonenumbers
email-validator
aiosmtplib
python-multipart
openpyxl
alembic
geoalchemy2
python-jose[cryptography]
passlib[bcrypt]
python-dateutil
redis
